Fixed undo bug
When a move was undone, and it was played again, the eval immediately
returned the eval froom hash without setting the next move. Now this
was no problem if only one move was undone, because it remained in
the global variable, but when multiple, it made a move that was
basically memory junk in that pos.