Written the move ordering function
not implemented yet, as it may needs some xtra care.
There are a lot of collisions in the hash table, which is,
at the moment 1 Mb. I tried varying the seed, and it does
seem like the case is slightly nicer, but we need to be
careful anyways. The logic was a bit messy at first because
of this not using a pure minimax search, but maybe now it is right.