Fixed fen input bug
It never read the halfmove and the move counter, because i forgot how sscanf works :((

Toggled one - to + in the engine, because it may have misinterpreted some positions, where there are multiple mates with different lengths, and it wouldn't found the shortest.

Changed the formatting of printmove to represent the more often used format with no spaces, and only lowercase letters.
e.g. e7 e8 Q => e7e8q